Refactor: Break out Login Process
